Fixed a couple of error messages, made slight change to tch.h

+++ PioneerLSeries.cydsn/structs.h
@@ -7,19 +7,19 @@ This file contains the data structs for the robotics arm motor board
 */
 #ifndef MY_STRUCTS
 #define MY_STRUCTS
-//#include "top.h"
+    
+#include "constants.h"
     
 //Potentiometer data struct
 typedef struct
 {
-    int16 potRaw[NUM_MOTORS];
-    //int potRaw[NUM_MOTORS];
+    int potRaw[NUM_MOTORS];
 } potData;
 
 //LCD Display data for debug purposes
 typedef struct
 {
-    int16 *potRawPtr[NUM_MOTORS];
+    int *potRawPtr[NUM_MOTORS];
 } displayData;
 
 #endif
